requirements.txt
WeasyPrint==50
Adafruit-Blinka==2.5.0
adafruit-circuitpython-74hc595==1.0.2
adafruit-circuitpython-busdevice==4.0.0
adafruit-circuitpython-charlcd==3.2.1
adafruit-circuitpython-mcp230xx==2.2.0
Adafruit-PlatformDetect==1.3.4
Adafruit-PureIO==0.2.3
boost==0.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cssselect2==0.2.2
decorator==4.4.0
defusedxml==0.6.0
gattlib==0.20150805
html5lib==1.0.1
idna==2.8
inky==0.0.5
Mastodon.py==1.4.6
numpy==1.17.2
picamera==1.13
Pillow==6.2.0
pkg-resources==0.0.0
PyBluez==0.22
pycparser==2.19
Pyphen==0.9.5
python-dateutil==2.8.0
python-magic==0.4.15
python-pushover==0.4
pytz==2019.2
requests==2.22.0
rpi-ws281x==4.2.2
RPi.GPIO==0.7.0
six==1.12.0
smbus2==0.3.0
spidev==3.4
SQLAlchemy==1.3.8
sysv-ipc==1.0.0
tinycss2==1.0.2
urllib3==1.25.6
webencodings==0.5.1
